+++ src/main/java/com/lxp/aplus/category/presentation/response/CategoryResponse.java
@@ -1,7 +1,7 @@
 package com.lxp.aplus.category.presentation.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.lxp.aplus.category.application.result.CategoryResult;
+import com.lxp.aplus.category.application.dto.response.CategoryResult;
 import lombok.Builder;
 
 import java.util.List;
